+++ components/src/core/dropdown-toolbar/dropdown-toolbar.component.ts
@@ -3,26 +3,26 @@ import { isEmptyView } from '../../utils/utils';
 import { MatMenuTrigger } from '@angular/material/menu';
 
 @Component({
-    selector: 'pxb-selection-toolbar',
+    selector: 'pxb-dropdown-toolbar',
     encapsulation: ViewEncapsulation.None,
     template: `
-        <mat-toolbar fxLayout="row" class="pxb-selection-toolbar">
+        <mat-toolbar fxLayout="row" class="pxb-dropdown-toolbar">
             <div
                 #icon
-                class="pxb-selection-toolbar-icon-wrapper"
-                [class.pxb-selection-toolbar-no-icon]="isIconEmpty(iconEl)"
+                class="pxb-dropdown-toolbar-icon-wrapper"
+                [class.pxb-dropdown-toolbar-no-icon]="isIconEmpty(iconEl)"
             >
                 <ng-content select="[pxb-nav-icon]"></ng-content>
             </div>
-            <div *ngIf="title" class="pxb-selection-toolbar-text-content-container">
-                <div class="pxb-selection-toolbar-title mat-title">{{ title }}</div>
+            <div *ngIf="title" class="pxb-dropdown-toolbar-text-content-container">
+                <div class="pxb-dropdown-toolbar-title mat-title">{{ title }}</div>
                 <div
                     *ngIf="subtitle"
-                    class="pxb-selection-toolbar-subtitle-container mat-subheading-2"
-                    [matMenuTriggerFor]="selectionToolbarMenu"
+                    class="pxb-dropdown-toolbar-subtitle-container mat-subheading-2"
+                    [matMenuTriggerFor]="dropdownToolbarMenu"
                     #menuTrigger="matMenuTrigger"
                 >
-                    <span class="pxb-selection-toolbar-subtitle mat-subheading-2">{{ subtitle }}</span>
+                    <span class="pxb-dropdown-toolbar-subtitle mat-subheading-2">{{ subtitle }}</span>
                     <mat-icon class="dropdown-arrow" [class.rotated-dropdown-arrow]="menuTrigger.menuOpen"
                         >arrow_drop_down</mat-icon
                     >
@@ -30,18 +30,14 @@ import { MatMenuTrigger } from '@angular/material/menu';
             </div>
             <pxb-spacer></pxb-spacer>
             <ng-content></ng-content>
-            <mat-menu
-                #selectionToolbarMenu="matMenu"
-                [overlapTrigger]="false"
-                class="pxb-selection-toolbar-menu-wrapper"
-            >
+            <mat-menu #dropdownToolbarMenu="matMenu" [overlapTrigger]="false" class="pxb-dropdown-toolbar-menu-wrapper">
                 <ng-content select="[pxb-toolbar-menu]"></ng-content>
             </mat-menu>
         </mat-toolbar>
     `,
-    styleUrls: ['./selection-toolbar.component.scss'],
+    styleUrls: ['./dropdown-toolbar.component.scss'],
 })
-export class SelectionToolbarComponent {
+export class DropdownToolbarComponent {
     @Input() title: string;
     @Input() subtitle: string;
     @ViewChild('icon', { static: true }) iconEl: ElementRef;