+++ components/src/core/selection-toolbar/public-api.ts
@@ -1,2 +1,0 @@
-export * from './selection-toolbar.component';
-export * from './selection-toolbar.module';