+++ demos/storybook/stories/dropdown-toolbar/with-basic-usage.stories.ts
@@ -2,13 +2,13 @@ import { text } from '@storybook/addon-knobs';
 
 export const withBasicUsage = (): any => ({
     template: `
-       <pxb-selection-toolbar [title]="title" [subtitle]="state.selected || subtitle">
+       <pxb-dropdown-toolbar [title]="title" [subtitle]="state.selected || subtitle">
         <ng-container pxb-toolbar-menu>
             <button mat-menu-item (click)="updateSubtitle('Test Item 1', state)">Test Item 1</button>
             <button mat-menu-item (click)="updateSubtitle('Test Item 2', state)">Test Item 2</button>
             <button mat-menu-item (click)="updateSubtitle('Test Item 3', state)">Test Item 3</button>
         </ng-container>
-       </pxb-selection-toolbar>
+       </pxb-dropdown-toolbar>
     `,
     props: {
         title: text('title', 'Title'),