+++ demos/storybook/stories/dropdown-toolbar/with-custom-menu.stories.ts
@@ -4,7 +4,7 @@ import { action } from '@storybook/addon-actions';
 
 export const withCustomMenu = (): any => ({
     template: `
-       <pxb-selection-toolbar [title]="title" [subtitle]="subtitle">
+       <pxb-dropdown-toolbar [title]="title" [subtitle]="subtitle">
         <ng-container pxb-toolbar-menu>
             <mat-nav-list>
                 <pxb-info-list-item (click)="click()"><mat-icon [style.color]="colors.blue[500]" pxb-icon>business</mat-icon><span pxb-title>Atlanta</span></pxb-info-list-item>
@@ -12,7 +12,7 @@ export const withCustomMenu = (): any => ({
                 <pxb-info-list-item (click)="click()"><mat-icon [style.color]="colors.blue[500]" pxb-icon>apartment</mat-icon><span pxb-title>New York</span></pxb-info-list-item>
             </mat-nav-list>
         </ng-container>
-       </pxb-selection-toolbar>
+       </pxb-dropdown-toolbar>
     `,
     props: {
         title: text('title', 'Title'),